refactor: replace old create experiment modal with new modal
