+++ webui/react/src/components/CreateExperimentModal.module.scss
@@ -1,3 +1,3 @@
 .error {
-  margin: var(--theme-sizes-layout-medium) var(--theme-sizes-layout-large);
+  margin: var(--theme-sizes-layout-large);
 }
